requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
dataclasses==0.6
deprecation==2.1.0
Faker==15.3.3
firebase-admin==5.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
fonttools==4.38.0
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gotrue==0.2.0
greenlet==1.1.3
grpcio==1.48.1
grpcio-status==1.48.1
h11==0.13.0
httpcore==0.12.3
httplib2==0.20.4
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pluggy==1.0.0
postgrest-py==0.4.0
proto-plus==1.22.1
protobuf==4.21.6
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
websockets==9.1
Werkzeug==2.2.2
